rename and create separate modules for search
