--- src/datagen/requirements.txt
@@ -1,8 +1,0 @@
-watermark 
-python-snappy 
-pandas  
-matplotlib 
-sentence_transformers 
-tqdm 
-numpy
-pyarrow
