+++ src/model/generate_training_data.sql
